tws fix

+++ MBAlertView/Views/MBAlertViewButton.m
@@ -56,7 +56,8 @@
     [[self colorForButtonStyle] setFill];   
     [path fill];
     
-    float actualSize = 0;
+    // fix font size
+    CGFloat actualSize = 0;
     [_title sizeWithFont:kButtonFont minFontSize:8 actualFontSize:&actualSize forWidth:self.bounds.size.width - 20 lineBreakMode:NSLineBreakByClipping];
     CGSize otherSize = [_title sizeWithFont:[UIFont boldSystemFontOfSize:actualSize]];
